+++ sdk/resourcemanager/datafactory/armdatafactory/models.go
@@ -328,18 +328,78 @@ func (a *AmazonMWSSource) GetTabularSource() *TabularSource {
 
 // AmazonRdsForLinkedServiceTypeProperties - AmazonRdsForOracle database linked service properties.
 type AmazonRdsForLinkedServiceTypeProperties struct {
-	// REQUIRED; The connection string. Type: string, SecureString or AzureKeyVaultSecretReference.
+	// Authentication type for connecting to the AmazonRdsForOracle database. Only used for Version 2.0.
+	AuthenticationType *AmazonRdsForOracleAuthenticationType
+
+	// The connection string. Type: string, SecureString or AzureKeyVaultSecretReference. Only used for Version 1.0.
 	ConnectionString any
+
+	// Specifies the desired data integrity behavior when this client connects to a server. Supported values are accepted, rejected,
+	// requested or required, default value is required. Type: string. Only used
+	// for Version 2.0.
+	CryptoChecksumClient any
+
+	// Specifies the crypto-checksum algorithms that client can use. Supported values are SHA1, SHA256, SHA384, SHA512, default
+	// value is (SHA512). Type: string. Only used for Version 2.0.
+	CryptoChecksumTypesClient any
+
+	// Specifies whether to use bulk copy or batch insert when loading data into the database, default value is true. Type: boolean.
+	// Only used for Version 2.0.
+	EnableBulkLoad any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
 
+	// Specifies the encryption client behavior. Supported values are accepted, rejected, requested or required, default value
+	// is required. Type: string. Only used for Version 2.0.
+	EncryptionClient any
+
+	// Specifies the encryption algorithms that client can use. Supported values are AES128, AES192, AES256, 3DES112, 3DES168,
+	// default value is (AES256). Type: string. Only used for Version 2.0.
+	EncryptionTypesClient any
+
+	// Specifies the number of bytes that the driver allocates to fetch the data in one database round-trip, default value is
+	// 10485760. Type: integer. Only used for Version 2.0.
+	FetchSize any
+
+	// Specifies whether the driver returns column value with the TIMESTAMP WITH TIME ZONE data type as DateTime or string. This
+	// setting is ignored if supportV1DataTypes is not true, default value is true.
+	// Type: boolean. Only used for Version 2.0.
+	FetchTswtzAsTimestamp any
+
+	// Specifies the amount that the source initially fetches for LOB columns, default value is 0. Type: integer. Only used for
+	// Version 2.0.
+	InitialLobFetchSize any
+
+	// Specifies a command that is issued immediately after connecting to the database to manage session settings. Type: string.
+	// Only used for Version 2.0.
+	InitializationString any
+
 	// The Azure key vault secret reference of password in connection string.
 	Password SecretBaseClassification
+
+	// The location of AmazonRdsForOracle database you want to connect to, the supported forms include connector descriptor, Easy
+	// Connect (Plus) Naming and Oracle Net Services Name (Only self-hosted IR).
+	// Type: string. Only used for Version 2.0.
+	Server any
+
+	// Specifies the number of cursors or statements to be cached for each database connection, default value is 0. Type: integer.
+	// Only used for Version 2.0.
+	StatementCacheSize any
+
+	// Specifies whether to use the Version 1.0 data type mappings. Do not set this to true unless you want to keep backward compatibility
+	// with Version 1.0's data type mappings, default value is false. Type:
+	// boolean. Only used for Version 2.0.
+	SupportV1DataTypes any
+
+	// The AmazonRdsForOracle database username. Type: string. Only used for Version 2.0.
+	Username any
 }
 
-// AmazonRdsForOracleLinkedService - AmazonRdsForOracle database.
+// AmazonRdsForOracleLinkedService - AmazonRdsForOracle database. This linked service has supported version property. The
+// Version 1.0 is scheduled for deprecation while your pipeline will continue to run after EOL but without any bug fix
+// or new features.
 type AmazonRdsForOracleLinkedService struct {
 	// REQUIRED; Type of linked service.
 	Type *string
@@ -3502,6 +3562,9 @@ type AzureDatabricksLinkedServiceTypeProperties struct {
 
 	// The credential reference containing authentication information.
 	Credential *CredentialReference
+
+	// The data security mode for the Databricks Cluster. Type: string (or Expression with resultType string).
+	DataSecurityMode any
 
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
@@ -9272,6 +9335,83 @@ type DataMapperMapping struct {
 
 	// Name of the target table
 	TargetEntityName *string
+}
+
+// DatabricksJobActivity - Databricks Job activity.
+type DatabricksJobActivity struct {
+	// REQUIRED; Activity name.
+	Name *string
+
+	// REQUIRED; Type of activity.
+	Type *string
+
+	// REQUIRED; Databricks Job activity properties.
+	TypeProperties *DatabricksJobActivityTypeProperties
+
+	// OPTIONAL; Contains additional key/value pairs not defined in the schema.
+	AdditionalProperties map[string]any
+
+	// Activity depends on condition.
+	DependsOn []*ActivityDependency
+
+	// Activity description.
+	Description *string
+
+	// Linked service reference.
+	LinkedServiceName *LinkedServiceReference
+
+	// Status result of the activity when the state is set to Inactive. This is an optional property and if not provided when
+	// the activity is inactive, the status will be Succeeded by default.
+	OnInactiveMarkAs *ActivityOnInactiveMarkAs
+
+	// Activity policy.
+	Policy *ActivityPolicy
+
+	// Activity state. This is an optional property and if not provided, the state will be Active by default.
+	State *ActivityState
+
+	// Activity user properties.
+	UserProperties []*UserProperty
+}
+
+// GetActivity implements the ActivityClassification interface for type DatabricksJobActivity.
+func (d *DatabricksJobActivity) GetActivity() *Activity {
+	return &Activity{
+		AdditionalProperties: d.AdditionalProperties,
+		DependsOn:            d.DependsOn,
+		Description:          d.Description,
+		Name:                 d.Name,
+		OnInactiveMarkAs:     d.OnInactiveMarkAs,
+		State:                d.State,
+		Type:                 d.Type,
+		UserProperties:       d.UserProperties,
+	}
+}
+
+// GetExecutionActivity implements the ExecutionActivityClassification interface for type DatabricksJobActivity.
+func (d *DatabricksJobActivity) GetExecutionActivity() *ExecutionActivity {
+	return &ExecutionActivity{
+		AdditionalProperties: d.AdditionalProperties,
+		DependsOn:            d.DependsOn,
+		Description:          d.Description,
+		LinkedServiceName:    d.LinkedServiceName,
+		Name:                 d.Name,
+		OnInactiveMarkAs:     d.OnInactiveMarkAs,
+		Policy:               d.Policy,
+		State:                d.State,
+		Type:                 d.Type,
+		UserProperties:       d.UserProperties,
+	}
+}
+
+// DatabricksJobActivityTypeProperties - Databricks Job activity properties.
+type DatabricksJobActivityTypeProperties struct {
+	// REQUIRED; The Id of the Databricks Job to be executed. Type: string (or Expression with resultType string).
+	JobID any
+
+	// Job parameters to be used for each run of this job. If the job takes a parameter that is not specified, the default value
+	// from the job will be used.
+	JobParameters map[string]any
 }
 
 // DatabricksNotebookActivity - DatabricksNotebook activity.
@@ -16239,6 +16379,10 @@ type ImpalaLinkedServiceTypeProperties struct {
 	// Specifies whether the connections to the server are encrypted using SSL. The default value is false.
 	EnableSSL any
 
+	// Specify whether to enable server SSL certificate validation when you connect.Always use System Trust Store (for V2 only).
+	// The default value is true.
+	EnableServerCertificateValidation any
+
 	// The encrypted credential used for authentication. Credentials are encrypted using the integration runtime credential manager.
 	// Type: string.
 	EncryptedCredential *string
@@ -16248,6 +16392,9 @@ type ImpalaLinkedServiceTypeProperties struct {
 
 	// The TCP port that the Impala server uses to listen for client connections. The default value is 21050.
 	Port any
+
+	// The transport protocol to use in the Thrift layer (for V2 only). Default value is Binary.
+	ThriftTransportProtocol *ImpalaThriftTransportProtocol
 
 	// The full path of the .pem file containing trusted CA certificates for verifying the server when connecting over SSL. This
 	// property can only be set when using SSL on self-hosted IR. The default value